Safari-611.1.10/LayoutTests/fast/dom/DeviceOrientation/updates-expected.txt

Tests that updates to the orientation causes new events to fire.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deviceOrientationEvent.alpha is mockEvent.alpha
PASS deviceOrientationEvent.beta is mockEvent.beta
PASS deviceOrientationEvent.gamma is mockEvent.gamma
PASS deviceOrientationEvent.alpha is mockEvent.alpha
PASS deviceOrientationEvent.beta is mockEvent.beta
PASS deviceOrientationEvent.gamma is mockEvent.gamma
PASS successfullyParsed is true

TEST COMPLETE
